🐛 fix: split ScenarioContentValidatorTests to satisfy type_body_length
The new per-field API tests pushed the suite struct body to 308 lines —
swiftlint's `type_body_length` cap is 250. Move the per-field tests
into a sibling file `ScenarioContentValidatorTests+PerField.swift` as
an extension of the suite struct, per testing.md's split convention.
The split file does NOT declare a new `@Suite`; it inherits
`.timeLimit(.minutes(1))` and `@MainActor` from the primary suite
declaration. Same-suite execution avoids the parallel-suite race
testing.md warns about.
